% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/eqmcc_args.R
\name{eqmcc_args}
\alias{eqmcc_args}
\title{Get \code{eqmcc()} arguments}
\usage{
eqmcc_args()
}
\value{
Returns a named list of arguments taken by \code{eqmcc()}
}
\description{
\code{eqmcc_args} creates an empty list of arguments taken
by \code{eqmcc()}
}
\examples{
a<-eqmcc_args()
}

