% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_qca_datasets.R
\name{get_qca_datasets}
\alias{get_qca_datasets}
\title{Get QCA datasets}
\usage{
get_qca_datasets(csQCA, mvQCA, fsQCA)
}
\arguments{
\item{csQCA}{logical; if TRUE, download all available csQCA datasets}

\item{mvQCA}{logical; if TRUE, download all available mvQCA datasets}

\item{fsQCA}{logical; if TRUE, download all available fsQCA datasets}
}
\value{
Returns a named list of QCA datasets with author identifiters
}
\description{
\code{get_qca_datasets} downloads a list of QCA datasets from the
COMPASSS.org database
}
\examples{
# Get all csQCA datasets (not run)

# a<-get_qca_datasets(csQCA = TRUE)

# Get all csQCA and fsQCA datasets (not run)

# a<-get_qca_datasets(csQCA = TRUE, fsQCA = TRUE)
}

