% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hh.R
\docType{data}
\name{hh}
\alias{hh}
\title{Replication from Hussain & Howard}
\format{A data frame with 20 rows and 11 variables}
\source{
\url{http://philhoward.org/wp-content/uploads/2012/11/International-Studies-Review-Replication-Data.csv}
}
\usage{
hh
}
\description{
A dataset containing fuzzy-set membership scores for 20 countries
in several sets, including average income, wealth inequality, level of unemployment,
population age, digital connectivity, censorship sophistication, fuel dependence of
the economy, regime fragility, and social movement success.
}
\references{
Muzammil M. Hussain and Philip N. Howard. 2013.
"What Best Explains Successful Protest Cascades? ICTs and the Fuzzy Causes of the Arab Spring."
International Studies Review (15) 1, pp 48-66.
}
\keyword{datasets}

