% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/hicks_20.R
\docType{data}
\name{hicks_20}
\alias{hicks_20}
\title{Replication Data from Hicks et al. (1995)}
\format{A data frame with 15 rows and 7 variables}
\source{
\url{http://www.compasss.org/bibliography/data/HicksEtal1995set1.csv}
}
\usage{
hicks_20
}
\description{
A dataset containing crisp-set membership scores for 15 countries in 1920
in several sets, including patriarchal statism, unitary democracy, working-
class mobilization, liberal government, catholic government, and the consolidation
of income security programs.
}
\references{
Hicks, Alexander, Joya Misra, and Tang Nah Ng. 1995.
"The Programmatic Emergence of the Social Security State."
American Sociological Review 60 (3):329-49.
}
\keyword{datasets}

