% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rfca_args.R
\name{rfca_args}
\alias{rfca_args}
\title{Get \code{rfca()} arguments}
\usage{
rfca_args()
}
\value{
Returns a named list of arguments taken by \code{rfca()}
}
\description{
\code{rfca_args} creates an empty list of arguments taken
by \code{rfca()}
}
\examples{
a<-rfca_args()
}

